Replace .status files by .stamp files, fix logging
An existing .stamp file now indicates a successful build. I may add
support for dependency resolution (a.k.a. A depends on B; if we rebuild
B, then we also force a rebuild of A), but that's secondary for the
moment.

Logs were falsely cleared at the beginning of the build, even if the
build was skipped.