Revert run script to python2, fix paths to various files
If running with gdb, the container is launched with tmux, in order to
get one prompt for gdb and one prompt for the qemu console each (via
Unix socket).

If running without gdb, there is currently no way to interact with the
VM (stdio console prevents qemu from forking into the background, and
the abovementioned tmux method is not usable when spawning multiple
instances). The goal is to provide the console via telnet.