Split data and chef directory for docker, fix build script
By default, /var/local/chef is shared to the container's `/data`,
containing both the VM and the output data location ("expdata").

Qemu is supposed to be installed + a few files are supposed to be copied
into the build directory's `opt`. It remains a mystery how it could work
before this was done (and more why it suddenly stopped working).