ccli: Attempt to solve the ACL issue
Even if setting the default ACL access rights for both users
(431 and $(id -u)) to rwx, the resulting files are still r-x, causing
problems at clean-up (currently requires root).

This is a desperate attempt to make sure that the ACLs are right.
According to non-Docker tests, this works. With Docker, it doesn't.