run.py: Fix remainings of python2, add missing --debug flag handling
IT RUNS! It's probably sane to add a 'troubleshooting' section/file
somewhere. The amount of things to consider (ACLs, existence of
files/directories, the right arguments to be passed (and whether paths
are in the context of the container or the host), ...) is horribly long.

Add example config.yml. The file structure in the main directory is a
big mess, it's likely to be cleaned up soon.