ccli: Restructure scripts and their locations
The build script is now a subcommand of ccli.py. Unlike the other
subcommands of ccli.py however, it can still be invoked stand-alone.

The half-baked python scripts in the 'batch' directory have been
removed, as they are currently not necessary (and it would require a
major rewrite of *everything*, which is clearly out of scope).