ccli: Add clean script
When running `mkdir -p` on a host-shared directory from within docker,
the ACLs do not seem to be respected. I don't know whether this is a bug
or something I missed. In any case, a user without root permissions is
no longer able to clean build files, but needs to pass through docker,
which may be tedious at times.