ccli: Commits for (aborted) debootstrap
Bootloader installation becomes a minor PITA if we are not in the system
itself. We could of course manually tweak (read: sed) the configuration
files so that GRUB finds the right disk/partition, but the probability
that something goes wrong along the way is not worth the trouble.

Now going for providiing premade disk images.