ccli: vm: Bundle ISO image in download
The .tar.gz should ideally be produced with:

    tar cvzf <VM name>.tar.gz <VM name>.qcow2 <ISO image>.iso

where <ISO image> is the associated ISO image used to create the VM. It
is currently hardcoded in the script. If we start providing multiple
images, we may want to set up some online archive that holds these kinds
of information (to simplify updating them), but for now this works.