ccli: Change paradigm: run.py calls itself
The problem is that GNU parallel takes stdin as input, but we need to
execute groups of two processes in parallel (launching qemu and
connecting to the command watchdog).
In order to be able to use GNU parallel for that, we'd need a separate
'connect' option in the run script. It's probably wise to add it in the
long run.

Usage:
	./run sym --batch-file 'foo.yml' --out-dir 'bar' | parallel

This commit was originally committed on 2015-05-26 09:21 by ayekat and
has been partially cherry-picked.