Revert "Make guest submodule publicly available"
This reverts commit 3805623c87d424e69584974ef00d7602bf2efc07.
Repositories *can* be cloned via SSH even by non-collaborators. The only
people left out now are those without a Github account. As S²E assumes
that most of its users have a Github account, the issue can be resolved
through a note in the documentation for how to clone without an account.